GitLab CI: speed up execution (#5235)
* gitlab-ci: add collapsable sections (with recorded durations)

https://docs.gitlab.com/ci/jobs/job_logs/#custom-collapsible-sections

* gitlab-ci: increase test threads to 4

* gitlab-ci: standalone tests only need one test thread

* gitlab-ci: use worksteal distribution more often

so a single core is not burdened with all tests at the end

* wdl tests: use pytest fixtures to support parallel execution

* gitlab-ci: run conformance tests in parallel

No need to run the conformance tests again in "cwl-v1.2" if we are
running the badge generation

* gitlab-ci: enable caching of pip & mypy

* gitlab-ci: adjust sections

* sort inputs

* lib/docker: add typing, switch to pytest from unittest

* test/src/busTest.py: add typing, switch to pytest from unittest

* test/src/deferredFunctionTest.py: add typing, switch to pytest from unittest

* test/src/importExportFileTest.py: add typing, switch to pytest from unittest

* test/src/jobEncapsulationTest.py: add typing, switch to pytest from unittest

* test/src/jobServiceTest.py: add typing, switch to pytest from unittest

* test/src/jobTest.py: add typing, switch to pytest from unittest

* test/src/miscTests.py: add typing, switch to pytest from unittest

* test/src/regularLogTest.py: add typing, switch to pytest from unittest

* test_iam: some tests need real AWS

* Makefile: organize pytest options; add overrides for logging, durations, and verbosity

fixed common pytest options have been moved to setup.cfg

* pytest marks: make safe for direct pytest.mark notation

* cwlTest: add missing marks

* ToilTest: remove unused parts; simplify others

* test skip online: unify

* pytest: default to worksteal test distribution

* Deprecate ToilTest

* test/utils/utilsTest.py: add typing, switch to pytest from unittest

* test/src/systemTest.py: add typing, switch to pytest from unittest

* test/src/resumabilityTest.py: add typing, switch to pytest from unittest

* test/src/retainTempDirTest.py: add typing, switch to pytest from unittest

* test/src/threadingTest.py: add typing, switch to pytest from unittest

* test/src/restartDAGTest.py: add typing, switch to pytest from unittest

* test/utils/toilDebugTest.py: add typing, switch to pytest from unittest

* test/lib/test_history.py: add typing, switch to pytest from unittest

* test/src/promisedRequirementTest.py: add typing, switch to pytest from unittest

* test/src/resourceTest.py: add typing, switch to pytest from unittest

+++ src/toil/batchSystems/cleanup_support.py
@@ -45,7 +45,7 @@ class BatchSystemCleanupSupport(BatchSystemLocalSupport):
         return contexts
 
     def __init__(
-        self, config: Config, maxCores: float, maxMemory: int, maxDisk: int
+        self, config: Config, maxCores: float, maxMemory: float, maxDisk: int
     ) -> None:
         super().__init__(config, maxCores, maxMemory, maxDisk)
